feat: Implement comment count aggregation by post IDs in CommentRepositoryCustom
- Added a method to count comments by post IDs using an IN query combined with GROUP BY, adhering to Instagram's policy of counting only root comments (parent is null).
- Updated PostService to utilize the new comment counting method for efficient data retrieval in the main feed.
- Modified PostResponse to reflect the change in comment count data type from int to long for consistency.
- Enhanced PostMapper to accommodate the new comment count parameter in response construction.
- Updated unit tests to validate the new comment counting logic and ensure accurate integration with the feed response.

+++ src/main/java/com/example/instagramclone/domain/post/api/PostResponse.java
@@ -16,7 +16,7 @@ public record PostResponse(
         List<PostImageResponse> images,
         LocalDateTime createdAt,
         LikeStatusResponse likeStatus,
-        int commentCount
+        long commentCount
 ) {
     
 }
